backend/requirements.txt
fastapi
uvicorn[standard]
ollama
python-dotenv
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
httpx
